support weight size bigger than 2g

+++ inc/graph/utils/tensor_utils.h
@@ -38,12 +38,12 @@ class TensorUtils {
   static void CopyTensor(const GeTensor &from, GeTensor &to);
   static ge::graphStatus GetSize(const GeTensorDesc &tensor_desc, int64_t &size);
   static void SetSize(GeTensorDesc &tensor_desc, const int64_t size);
-  static uint32_t GetWeightSize(const ConstGeTensorPtr &tensor_ptr);
-  static uint32_t GetWeightSize(const GeTensor &tensor);
-  static uint32_t GetWeightSize(const GeTensorDesc &tensor_desc);
+  static int64_t GetWeightSize(const ConstGeTensorPtr &tensor_ptr);
+  static int64_t GetWeightSize(const GeTensor &tensor);
+  static int64_t GetWeightSize(const GeTensorDesc &tensor_desc);
   static uint8_t *GetWeightAddr(const ConstGeTensorPtr &tensor_ptr, const uint8_t *const base);
   static uint8_t *GetWeightAddr(const GeTensor &tensor, const uint8_t *const base);
-  static void SetWeightSize(GeTensorDesc &tensor_desc, const uint32_t size);
+  static void SetWeightSize(GeTensorDesc &tensor_desc, const int64_t size);
   static ge::graphStatus GetReuseInput(const GeTensorDesc &tensor_desc, bool &flag);
   static void SetReuseInput(GeTensorDesc &tensor_desc, const bool flag);
   static ge::graphStatus GetOutputTensor(const GeTensorDesc &tensor_desc, bool &flag);
